example: LlamaParseReader in SimpleDirectoryReader using parallel processing
