+++ examples/readers/package.json
@@ -11,7 +11,7 @@
     "start:pdf": "node --import tsx ./src/pdf.ts",
     "start:llamaparse": "node --import tsx ./src/llamaparse.ts",
     "start:notion": "node --import tsx ./src/notion.ts",
-    "start:llamaparse2": "node --import tsx ./src/llamaparse_2.ts"
+    "start:llamaparse2": "node --import tsx ./src/simple-directory-reader-with-llamaparse.ts"
   },
   "dependencies": {
     "llamaindex": "*"
